Add files via upload
The report sheds light upon “The economic crisis” started in the summer of 2007, though the
full impact was not felt till the bankruptcy of the investment bank, Lehmann Brothers
in September 2008. The next couple of years witnessed heavy job losses and
contraction in the GDP (Gross Domestic Product) of many countries in the West as
well as in the developing world. The objectives of the research were to assess the
human impacts of the crisis and to analyze whether responses by the government and
civil society were serving the interests of Indian citizens.